Fix controllers methods and add custom service exceptions to provide response statuses

+++ fitness-club/service/src/main/java/com/epam/fitness/service/impl/UserServiceImpl.java
@@ -7,6 +7,8 @@ import com.epam.fitness.entity.order.Order;
 import com.epam.fitness.entity.user.User;
 import com.epam.fitness.dao.api.UserDao;
 import com.epam.fitness.entity.UserDto;
+import com.epam.fitness.exception.EntityAlreadyExistsException;
+import com.epam.fitness.exception.EntityNotFoundException;
 import com.epam.fitness.exception.ServiceException;
 import com.epam.fitness.service.api.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,7 +37,7 @@ public class UserServiceImpl implements UserService {
     public List<UserDto> getClientsByTrainerId(int trainerId) throws ServiceException{
         Optional<User> trainerOptional = dao.findById(trainerId);
         User trainer = trainerOptional
-                .orElseThrow(() -> new ServiceException("Trainer with id " + trainerId + " not found!"));
+                .orElseThrow(() -> new EntityNotFoundException("Trainer with id " + trainerId + " not found!"));
         List<User> clients =  dao.findClientsOfTrainer(trainer);
         return mapper.mapToDto(clients);
     }
@@ -47,7 +49,8 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<UserDto> searchUsersByParameters(String firstName, String secondName, String email, SortOrder order)
+    public List<UserDto> searchUsersByParameters(String firstName, String secondName, String email,
+                                                 SortOrder order)
             throws ServiceException{
         List<User> users = dao.findUsersByParameters(firstName, secondName, email, order);
         return mapper.mapToDto(users);
@@ -57,7 +60,7 @@ public class UserServiceImpl implements UserService {
     public UserDto getUserByEmail(String email) throws ServiceException {
         Optional<User> userOptional = dao.findUserByEmail(email);
         User user = userOptional
-                .orElseThrow(() -> new ServiceException("User with email " + email + " not found!"));
+                .orElseThrow(() -> new EntityNotFoundException("User with email " + email + " not found!"));
         return mapper.mapToDto(user);
     }
 
@@ -72,39 +75,52 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void updateById(int id, UserDto userDto) throws ServiceException {
+    public UserDto updateById(int id, UserDto userDto) throws ServiceException {
         User user = mapper.mapToEntity(userDto);
         Optional<User> userOptional = dao.findById(id);
         User oldUser = userOptional
-                .orElseThrow(() -> new ServiceException("User with id " + id + " not found!"));
+                .orElseThrow(() -> new EntityNotFoundException("User with id " + id + " not found!"));
         List<Order> orders = oldUser.getOrders();
         user.setOrders(orders);
         String password = user.getPassword();
         String passwordHash = hashPassword(password);
         user.setPassword(passwordHash);
         user.setId(id);
-        dao.save(user);
+        User savedUser = dao.save(user);
+        return mapper.mapToDto(savedUser);
     }
 
     @Override
-    public void create(UserDto userDto) throws ServiceException{
+    public UserDto create(UserDto userDto) throws ServiceException{
+        Integer id = userDto.getId();
+        if(Objects.nonNull(id)){
+            Optional<User> userOptional = dao.findById(id);
+            if(userOptional.isPresent()){
+                throw new EntityAlreadyExistsException("User with id " + id + " already exists!");
+            }
+        }
         User user = mapper.mapToEntity(userDto);
     	String password = user.getPassword();
 		String passwordHash = hashPassword(password);
 		user.setPassword(passwordHash);
-        dao.save(user);
+		User savedUser = dao.save(user);
+        return mapper.mapToDto(savedUser);
     }
 
     @Override
     public UserDto getById(int id) throws ServiceException {
         Optional<User> userOptional = dao.findById(id);
         User user = userOptional
-                .orElseThrow(() -> new ServiceException("User with id " + id + " not found!"));
+                .orElseThrow(() -> new EntityNotFoundException("User with id " + id + " not found!"));
         return mapper.mapToDto(user);
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(int id) throws ServiceException {
+        Optional<User> userOptional = dao.findById(id);
+        if(!userOptional.isPresent()){
+            throw new EntityNotFoundException("User with id " + id + " not found!");
+        }
         dao.deleteById(id);
     }
